test(sdk): Split timeline tests into smaller modules
