+++ crates/matrix-sdk/src/room/timeline/tests/echo.rs
@@ -1,0 +1,152 @@
+use std::{io, sync::Arc};
+
+use assert_matches::assert_matches;
+use futures_signals::signal_vec::VecDiff;
+use futures_util::StreamExt;
+use matrix_sdk_test::async_test;
+use ruma::{
+    event_id,
+    events::{room::message::RoomMessageEventContent, AnyMessageLikeEventContent},
+};
+use serde_json::json;
+
+use super::{TestTimeline, ALICE, BOB};
+use crate::{
+    room::timeline::{event_item::EventSendState, EventTimelineItem},
+    Error,
+};
+
+#[async_test]
+async fn remote_echo_full_trip() {
+    let timeline = TestTimeline::new();
+    let mut stream = timeline.stream();
+
+    // Given a local event…
+    let txn_id = timeline
+        .handle_local_event(AnyMessageLikeEventContent::RoomMessage(
+            RoomMessageEventContent::text_plain("echo"),
+        ))
+        .await;
+
+    let _day_divider = assert_matches!(stream.next().await, Some(VecDiff::Push { value }) => value);
+
+    // Scenario 1: The local event has not been sent yet to the server.
+    {
+        let item = assert_matches!(stream.next().await, Some(VecDiff::Push { value }) => value);
+        let event = item.as_event().unwrap().as_local().unwrap();
+        assert_matches!(event.send_state, EventSendState::NotSentYet);
+    }
+
+    // Scenario 2: The local event has not been sent to the server successfully, it
+    // has failed. In this case, there is no event ID.
+    {
+        let some_io_error = Error::Io(io::Error::new(io::ErrorKind::Other, "this is a test"));
+        timeline.inner.update_event_send_state(
+            &txn_id,
+            EventSendState::SendingFailed { error: Arc::new(some_io_error) },
+        );
+
+        let item = assert_matches!(
+            stream.next().await,
+            Some(VecDiff::UpdateAt { value, index: 1 }) => value
+        );
+        let event = item.as_event().unwrap().as_local().unwrap();
+        assert_matches!(event.send_state, EventSendState::SendingFailed { .. });
+    }
+
+    // Scenario 3: The local event has been sent successfully to the server and an
+    // event ID has been received as part of the server's response.
+    let event_id = {
+        let event_id = event_id!("$W6mZSLWMmfuQQ9jhZWeTxFIM");
+
+        timeline.inner.update_event_send_state(
+            &txn_id,
+            EventSendState::Sent { event_id: event_id.to_owned() },
+        );
+
+        let item = assert_matches!(
+            stream.next().await,
+            Some(VecDiff::UpdateAt { value, index: 1 }) => value
+        );
+        let event = item.as_event().unwrap().as_local().unwrap();
+        assert_matches!(event.send_state, EventSendState::Sent { .. });
+
+        event_id
+    };
+
+    // Now, a sync has been run against the server, and an event with the same ID
+    // comes in.
+    timeline
+        .handle_live_custom_event(json!({
+            "content": {
+                "body": "echo",
+                "msgtype": "m.text",
+            },
+            "sender": &*ALICE,
+            "event_id": event_id,
+            "origin_server_ts": 5,
+            "type": "m.room.message",
+        }))
+        .await;
+
+    // The local echo is removed
+    assert_matches!(stream.next().await, Some(VecDiff::Pop { .. }));
+
+    // This day divider shouldn't be present, or the previous one should be
+    // removed. There being a two day dividers in a row is a bug, but it's
+    // non-trivial to fix and rare enough that we can fix it later (only happens
+    // when the first message on a given day is a local echo).
+    let _day_divider = assert_matches!(stream.next().await, Some(VecDiff::Push { value }) => value);
+
+    // … and the remote echo is added.
+    let item = assert_matches!(stream.next().await, Some(VecDiff::Push { value }) => value);
+    assert_matches!(item.as_event().unwrap(), EventTimelineItem::Remote(_));
+}
+
+#[async_test]
+async fn remote_echo_new_position() {
+    let timeline = TestTimeline::new();
+    let mut stream = timeline.stream();
+
+    // Given a local event…
+    let txn_id = timeline
+        .handle_local_event(AnyMessageLikeEventContent::RoomMessage(
+            RoomMessageEventContent::text_plain("echo"),
+        ))
+        .await;
+
+    let _day_divider = assert_matches!(stream.next().await, Some(VecDiff::Push { value }) => value);
+
+    let item = assert_matches!(stream.next().await, Some(VecDiff::Push { value }) => value);
+    let txn_id_from_event = item.as_event().unwrap().as_local().unwrap();
+    assert_eq!(txn_id, *txn_id_from_event.transaction_id);
+
+    // … and another event that comes back before the remote echo
+    timeline.handle_live_message_event(&BOB, RoomMessageEventContent::text_plain("test")).await;
+    let _day_divider = assert_matches!(stream.next().await, Some(VecDiff::Push { value }) => value);
+    let _bob_message = assert_matches!(stream.next().await, Some(VecDiff::Push { value }) => value);
+
+    // When the remote echo comes in…
+    timeline
+        .handle_live_custom_event(json!({
+            "content": {
+                "body": "echo",
+                "msgtype": "m.text",
+            },
+            "sender": &*ALICE,
+            "event_id": "$eeG0HA0FAZ37wP8kXlNkxx3I",
+            "origin_server_ts": 6,
+            "type": "m.room.message",
+            "unsigned": {
+                "transaction_id": txn_id,
+            },
+        }))
+        .await;
+
+    // … the local echo should be removed
+    assert_matches!(stream.next().await, Some(VecDiff::RemoveAt { index: 1 }));
+
+    // … and the remote echo added
+    let item = assert_matches!(stream.next().await, Some(VecDiff::Push { value }) => value);
+    assert_matches!(item.as_event().unwrap(), EventTimelineItem::Remote(_));
+}
